% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copci_dev_heatmap.R
\name{copci_dev_heatmap}
\alias{copci_dev_heatmap}
\title{Heatmap plot}
\usage{
copci_dev_heatmap(
  gene,
  image_data,
  norm_val = NULL,
  color_n = 75,
  colfun = colorRampPalette(brewer.pal(8, "YlOrRd"))
)
}
\arguments{
\item{gene}{A character of a gene name: CopciAB_ and a number. E.g., CopciAB_411205}

\item{image_data}{A list which is the output of the copci_image_data function}

\item{norm_val}{NULL or a character vector length of 2. If NULL the expression values will be normalized and centered between 0 and 1 based on the values of the selected gene. If it is a vector then the minimum and the maximum values should be added which will be used to normalize and center the data.}

\item{color_n}{Integer. Number of colors to pick from the color palette.}

\item{colfun}{Function. A color palette function.}
}
\description{
Heatmap plot of the expression data through the developmental stage of Coprinopsis cinerea
}
