% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copci_exp_plot.R
\name{copci_exp_plot}
\alias{copci_exp_plot}
\title{Boxplot of expression data}
\usage{
copci_exp_plot(
  gene,
  log = TRUE,
  expdata = exp_data,
  coldata = df_combined,
  order = "order",
  facetting = "plot_facets",
  color = "fill"
)
}
\arguments{
\item{gene}{A character of a gene name: CopciAB_ and a number. E.g., CopciAB_411205}

\item{log}{A logical indicating whether expression values should be log 2 transformed (Default is TRUE)}
}
\description{
Plotting expression data of Coprinopsis cinerea
}
