% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copci_image_data.R
\name{copci_image_data}
\alias{copci_image_data}
\title{Copci data image processing}
\usage{
copci_image_data(HyphalKnot = "H1_VM")
}
\arguments{
\item{HyphalKnot}{Two kind of Hyphal knot samples exist. You can choose of which expressiond data should be used: H1_VM or H1_cap}
}
\value{
A list of three elements. First element contains the image_df with three columns. The second element contains the image pointer created by image_read. The third element contains the expression data.
}
\description{
Creates a data based on Coprinopsis cinerea tissue-specific expression data that can be used for image heat map plotting.
}
