% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_data_process.R
\name{image_data_process}
\alias{image_data_process}
\title{Image data processing}
\usage{
image_data_process(image_df)
}
\arguments{
\item{image_df}{A dataframe of two columns. Column 1 should be named images. Column 2 should be named areas. Both columns should contain characters defining the images that will be imported, and the areas on the images that should be coloured.}

\item{coldata}{A dataframe containing infromation for the columns of the expression_data}
}
\value{
A list of two elements. First element contains the image_df with a third column defining the coordinates of the areas. The second element contains the image pointer created by image_read
}
\description{
Creates a data that can be used for image heat map plotting. It uses interactive image import area locator.
}
