% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_expression.R
\name{mean_expression}
\alias{mean_expression}
\title{Average expression}
\usage{
mean_expression(expression_data, coldata, sample_col, id_col)
}
\arguments{
\item{expression_data}{A matrix containing gene expressin data. Column names should be sample id and row names should be gene ids}

\item{coldata}{A dataframe containing infromation for the columns of the expression_data}

\item{sample_col}{Character. The name of the column in the coldata defining the sample id that can be found in the column nams of the expression_data}
}
\value{
A dataframe containing mean expression data with rownames of gene id and colnams of samples
}
\description{
Calculate the mean of expession data replicates
}
